UI: Separate UserProfileContainer
